Added Checkout and Check-in Methods
created checkout and check-in methods for books. when the methods are called they first check if the given ISBN is in the database, returns a message through the HTTP request if not. the methods then make sure the book isnt already checked out or in, depedning on which method is called and then updates the book's status accordingly. if successful, then the HTTP will receive a message saying so and if an error occurred, then the HTTP will receive a simple error message, providing some information as to why the request failed. any questions/concerns let me know on discord - Deryck
ps. I still have to add the whole "ask for the Card_id of the Borrower" to associate the checkout with them. so ill be working on that for right now.